Initialize NetworkCredential with string password
On my machine, the Password property was set automatically even
if we used the SecureString overload of the constructor.
According to the documentation, however, and on some other machines, the
Password property is set if using the (string,string) constructor.
Perhaps I got an newer version which did the conversion automatically...

The result of this was that the Password property always returned null,
which caused the download function to never be able to set authentication.

This commit will convert the secure string to a plain string before
we initialize the NetworkCredential object, making it work as expected.